Tailwind - v0. Bailing.
Ok - I've got tailwind working. But dang - it's a lot of
work to get back to a basically styled site.

Maybe the people who are fans of tailwind are trying to solve
a problem that isn't important to me.

I could *probably* jump start my work by eithe buying a framework
and template from someone like FlowBrite. But that costs 200-300!

Or I'm sure there are good free ones out there, but they are
not easy to find. And at the end of the day, I'd be back where
I am right now.

Sure bootstrap is old school, but it's easy, I know it, and I'm
done with the implementation.

I'm going back to it.